requirements.txt
accelerate==1.12.0
anyio==4.12.0
certifi==2025.11.12
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
denoising-diffusion-pytorch==2.2.5
einops==0.8.1
ema-pytorch==0.7.7
filelock==3.20.0
fonttools==4.61.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.7
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.6
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pillow==12.0.0
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
PyYAML==6.0.3
safetensors==0.7.0
scipy==1.16.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typer-slim==0.20.0
typing_extensions==4.15.0
